requirements.txt
asgiref==3.5.2
certifi==2022.5.18.1
cffi==1.15.0
chardet==4.0.0
crispy-tailwind==0.5.0
cryptography==37.0.2
Deprecated==1.2.13
dj-database-url==0.5.0
Django==3.2.14
django-admin-multiple-choice-list-filter==0.1.1
django-crispy-forms==1.14.0
django-debug-toolbar==3.4.0
django-on-heroku==1.1.2
gunicorn==20.1.0
idna==2.10
psycopg2-binary==2.9.3
pycparser==2.21
PyGithub==1.55
PyJWT==2.4.0
PyNaCl==1.5.0
python-decouple==3.6
pytz==2022.1
requests==2.25.1
sqlparse==0.4.2
urllib3==1.26.9
whitenoise==6.2.0
wrapt==1.14.1
django_filter==22.1
